refactor(cli/hotspots): split command into args/json/human submodules
Mirror the commands/summary split so the two largest CLI commands share one
structure. commands/hotspots.rs (1236 lines) → hotspots/ module:

- hotspots/mod.rs (294) — orchestration: HotspotsArgs, RankBy, run, run_inner,
  pattern-selection, and the test block.
- hotspots/json.rs (280) — JSON serialization (the *_to_json helpers).
- hotspots/human.rs (687) — human/table rendering (emit_human, findings/section
  tables, *_row, sort_*, format_bytes); internal helpers kept private.

Public surface (HotspotsArgs, RankBy, run) unchanged. Pure code-movement:
43/43 item-set parity, moved bodies byte-identical, CLI tests pass, clippy
clean, live hotspots OK.
